backend/requirements-langchain.txt
# Enhanced Requirements with LangChain Integration
# Core dependencies
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
python-docx==0.8.11
PyPDF2==3.0.1
azure-communication-email==1.0.0
sqlalchemy==2.0.23
qdrant-client==1.7.0
sentence-transformers==2.2.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9

# LangChain Core Dependencies
langchain==0.1.0
langchain-core==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
langchain-google-genai==0.0.6

# LangChain Memory and Storage
redis==5.0.1
chromadb==0.4.22

# LangChain Document Processing
pypdf==3.17.4
docx2txt==0.8
python-multipart==0.0.6
unstructured==0.11.8
pdfplumber==0.10.3

# LangChain Vector Stores and Embeddings
faiss-cpu==1.7.4
pinecone-client==2.2.4
weaviate-client==3.25.3

# LangChain Tools and Agents
duckduckgo-search==4.1.1
wikipedia==1.4.0
wolframalpha==5.0.0

# LangChain Callbacks and Monitoring
langsmith==0.0.69
tiktoken==0.5.2

# Enhanced Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78
Wand==0.6.13
pdf2image==1.16.3

# Async and Performance
aiohttp==3.9.1
asyncio-mqtt==0.16.1
celery==5.3.4
redis==5.0.1

# Monitoring and Logging
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk==1.40.4

# Testing and Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Additional AI/ML Libraries
transformers==4.36.2
torch==2.1.2
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2

# Web Scraping and Data Extraction
beautifulsoup4==4.12.2
selenium==4.15.2
requests-html==0.10.0

# API and Integration
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
httpx==0.25.2

# Security and Authentication
cryptography==41.0.8
passlib==1.7.4
python-jose==3.3.0

# Database and Caching
alembic==1.13.1
redis==5.0.1
memcached==1.59

# File Processing
python-magic==0.4.27
chardet==5.2.0
filetype==1.2.0
